py-api/pyproject.toml
[tool.poetry]
name = "py-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
weaviate-client = "^4.5.4"
langchain = "^0.1.14"
langchain-openai = "^0.1.1"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
kafka-python = "^2.0.2"
newsapi-python = "^0.2.7"
websockets = "^12.0"
redis = "^5.0.3"
langchain-weaviate = "^0.0.1"
lark = "^1.1.9"
six = "^1.16.0"
langchain-anthropic = "^0.1.11"


[build-system]
requires = ["poetry-core>=1.8.2"]
build-backend = "poetry.core.masonry.api"
